pyproject.toml
[tool.poetry]
name = "graphrag-mcp"
version = "0.1.0"
description = "GraphRAG MCP Server with hybrid search — no Neo4j required"
authors = ["you <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
graphrag = ">=2.0"
fastmcp = "0.2.0"
starlette = "^0.37"
uvicorn = "^0.29"
pandas = "^2.0"
pyarrow = "^15.0"
qdrant-client = "^1.9"
rank-bm25 = "^0.2"
pydantic-settings = "^2.0"
python-dotenv = "^1.0"
openai = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
ruff = "^0.4"
